% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbeta.R
\name{hbeta}
\alias{hbeta}
\title{Partial moments for the Beta Distribution}
\usage{
hbeta(x, order = 1, shape1, shape2, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of values \eqn{x}.}

\item{order}{the value of \eqn{k \in \{1,2\}}. Default is 1.}

\item{shape1, shape2}{non-negative parameters of the Beta distribution.}

\item{lower.tail}{logical; if TRUE (default), \eqn{\mathcal{A}} is
\eqn{X \leq x}, otherwise,  \eqn{\mathcal{A}} is \eqn{X \geq x}.}
}
\value{
Numeric value or vector with the moments.
}
\description{
Returns absolute partial moments of first and second
order, say

\deqn{\mathbb{E}(X^k \ \mathbb{I}\{X \in \mathcal{A}\}),}

with \eqn{k \in \{1,2\}} and \eqn{\mathcal{A}} either \eqn{X \leq x} or
\eqn{X \geq x}, for the Beta distribution with parameters \code{shape1}
and \code{shape2}.
}
\details{
The Beta distribution with parameters \code{shape1}
and \code{shape2} has density:

\deqn{
f(x)=\frac{\Gamma(a+b)}{\Gamma(a)\Gamma(b)}{x}^{a-1} {(1-x)}^{b-1}}

for \eqn{a > 0, b > 0} and \eqn{0 \le x \le 1} where the boundary values
at \eqn{x=0} or \eqn{x=1} are defined as by continuity (as limits).
}
