% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbinom.R
\name{hbinom}
\alias{hbinom}
\title{Partial moments for the Binomial Distribution}
\usage{
hbinom(x, order = 1, size, prob, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of values \eqn{x}.}

\item{order}{the value of \eqn{k \in \{1,2\}}. Default is 1.}

\item{size}{umber of trials (zero or more).}

\item{prob}{probability of success on each trial.}

\item{lower.tail}{logical: If TRUE (default), \eqn{\mathcal{A}} is
\eqn{X \leq x}; otherwise,  \eqn{\mathcal{A}} is \eqn{X \geq x}.}
}
\value{
Numeric value or vector with the moments.

If \code{size} is not an integer, NaN is returned.

For more details check \link\link{dbinom}.
}
\description{
Returns absolute partial moments of first and second
order, say

\deqn{\mathbb{E}(X^k \ \mathbb{I}\{X \in \mathcal{A}\}),}

with \eqn{k \in \{1,2\}} and \eqn{\mathcal{A}} either \eqn{X \leq x} or
\eqn{X \geq x}, for the Binomial distribution with parameters \code{size}
and \code{prob}.
}
\details{
The Binomial distribution with parameters \eqn{n=}\code{size}
and \eqn{p=}\code{prob} has density:

\deqn{p(x) = \binom{n}{x} p^x (1-p)^{n-x}}

for \eqn{x = 0, ..., n}.
}
