% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnbinom.R
\name{hnbinom}
\alias{hnbinom}
\title{Partial moments for the Negative Binomial Distribution}
\usage{
hnbinom(x, order = 1, size, prob, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of values \eqn{x}.}

\item{order}{the value of \eqn{k \in \{1,2\}}. Default is 1.}

\item{size}{target for number of successful trials, or dispersion parameter
(the shape parameter of the gamma mixing distribution).
Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}

\item{lower.tail}{logical: If TRUE (default), \eqn{\mathcal{A}} is
\eqn{X \leq x}; otherwise,  \eqn{\mathcal{A}} is \eqn{X \geq x}.}
}
\description{
Returns absolute partial moments of first and second
order, say

\deqn{\mathbb{E}(X^k \ \mathbb{I}\{X \in \mathcal{A}\}),}

with \eqn{k \in \{1,2\}} and \eqn{\mathcal{A}} either \eqn{X \leq x} or
\eqn{X \geq x}, for the Negative Binomial distribution with parameters
\code{size} and \code{prob}.
}
\details{
For more details check \link{dnbinom}.
}
