% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hweibull.R
\name{hweibull}
\alias{hweibull}
\title{Partial moments for the Weibull Distribution}
\usage{
hweibull(x, order, shape, scale = 1, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of values \eqn{x}.}

\item{order}{the value of \eqn{k \in \{1,2\}}. Default is 1.}

\item{shape, scale}{shape and scale parameters, the latter defaulting to 1.}

\item{lower.tail}{logical: If TRUE (default), \eqn{\mathcal{A}} is
\eqn{X \leq x}; otherwise,  \eqn{\mathcal{A}} is \eqn{X \geq x}.}
}
\value{
Numeric value or vector with the moments.
}
\description{
Returns absolute partial moments of first and second
order, say

\deqn{\mathbb{E}(X^k \ \mathbb{I}\{X \in \mathcal{A}\}),}

with \eqn{k \in \{1,2\}} and \eqn{\mathcal{A}} either \eqn{X \leq x} or
\eqn{X \geq x}, for the Weibull distribution with parameters \code{shape}
and \code{scale}.
}
\details{
For more details check \link{dweibull}.
}
