% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlnorm.R
\name{hlnorm}
\alias{hlnorm}
\title{Partial moments for the Lognormal Distribution}
\usage{
hlnorm(x, order = 1, meanlog = 0, sdlog = 1, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of values \eqn{x}.}

\item{order}{the value of \eqn{k \in \{1,2\}}. Default is 1.}

\item{meanlog, sdlog}{mean and standard deviation of the distribution on
the log scale with default values of 0 and 1 respectively.}

\item{lower.tail}{logical: If TRUE (default), \eqn{\mathcal{A}} is
\eqn{X \leq x}; otherwise,  \eqn{\mathcal{A}} is \eqn{X \geq x}.}
}
\value{
Numeric value or vector with the moments.
}
\description{
Returns absolute partial moments of first and second
order, say

\deqn{\mathbb{E}(X^k \ \mathbb{I}\{X \in \mathcal{A}\}),}

with \eqn{k \in \{1,2\}} and \eqn{\mathcal{A}} either \eqn{X \leq x} or
\eqn{X \geq x}, for the Lognormal distribution with parameters \code{meanlog}
and \code{sdlog}.
}
\details{
For more details check \link{dlnorm}.
}
