% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnorm.R
\name{hnorm}
\alias{hnorm}
\title{Partial moments for the Normal Distribution}
\usage{
hnorm(x, order = 1, mean = 0, sd = 1, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of values \eqn{x}.}

\item{order}{the value of \eqn{k \in \{1,2\}}. Default is 1.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}

\item{lower.tail}{logical: If TRUE (default), \eqn{\mathcal{A}} is
\eqn{X \leq x}; otherwise,  \eqn{\mathcal{A}} is \eqn{X \geq x}.}
}
\value{
Numeric value or vector with the moments.
}
\description{
Returns absolute partial moments of first and second
order, say

\deqn{\mathbb{E}(X^k \ \mathbb{I}\{X \in \mathcal{A}\}),}

with \eqn{k \in \{1,2\}} and \eqn{\mathcal{A}} either \eqn{X \leq x} or
\eqn{X \geq x}, for the Normal distribution with mean equal to
\code{mean} and standard deviation equal to \code{sd}.
}
\details{
For more details check \link{dnorm}.
}
