% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpois.R
\name{hpois}
\alias{hpois}
\title{Partial moments for the Poisson Distribution}
\usage{
hpois(x, order = 1, lambda, lower.tail = TRUE)
}
\arguments{
\item{x}{vector of values \eqn{x}.}

\item{order}{the value of \eqn{k \in \{1,2\}}. Default is 1.}

\item{lambda}{vector of (non-negative) means.}

\item{lower.tail}{logical: If TRUE (default), \eqn{\mathcal{A}} is
\eqn{X \leq x}; otherwise,  \eqn{\mathcal{A}} is \eqn{X \geq x}.}
}
\value{
Numeric value or vector with the moments.
}
\description{
Returns absolute partial moments of first and second
order, say

\deqn{\mathbb{E}(X^k \ \mathbb{I}\{X \in \mathcal{A}\}),}

with \eqn{k \in \{1,2\}} and \eqn{\mathcal{A}} either \eqn{X \leq x} or
\eqn{X \geq x}, for the Poisson distribution with parameter \code{lambda}.
}
\details{
For more details check \link{dpois}.
}
