% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexpec-package.R
\docType{package}
\name{pexpec-package}
\alias{pexpec}
\alias{pexpec-package}
\title{pexpec: Partial expectation computations for parametric distributions}
\description{
This package provides functions that calculate partial expectations for the most common parametric distribution models.
}
\author{
\strong{Maintainer}: Pedro Cosatto \email{pcosatto@fi.uba.ar} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
